requirements.txt
Flask
Flask-Login
flask-nav
argon2-cffi
psycopg2-binary
requests
sendgrid
